requirements.txt
anthropic>=0.42.0
openai>=1.50.0
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
pypdf>=5.0.0
pytest>=9.0.0
psycopg[binary]>=3.2.0
